Move QPDF inner class definitions to new QPDF_private.hh
