+++ libqpdf/QPDF_optimization.cc
@@ -2,7 +2,7 @@
 
 #include <qpdf/assert_debug.h>
 
-#include <qpdf/QPDF.hh>
+#include <qpdf/QPDF_private.hh>
 
 #include <qpdf/QPDFExc.hh>
 #include <qpdf/QPDFWriter_private.hh>
