+++ libqpdf/QPDF_pages.cc
@@ -1,4 +1,4 @@
-#include <qpdf/QPDF.hh>
+#include <qpdf/QPDF_private.hh>
 
 #include <qpdf/QPDFExc.hh>
 #include <qpdf/QTC.hh>
